Reorganization Class Creation and Completion
Reorgainized where the "AttackPhase" class was located so all phases could be compartmentalized within "PhaseProcessing" folder. Also created the "Phases" class that the other phases will inherit from and the "DraftPhase" and "FortifyPhase" classes that will be filled in next.

Created the "Regions" class to automate the creation of Region objects to assist in determining how many troops are owed come the "Draft" phase.

Created the "Troops" class to assist in troop modification (i.e. count, relocation, board assignment, etc.).